refactor service directory
