+++ src/main/java/com/example/AutoDocX/service/SessionManager.java
@@ -1,6 +1,6 @@
 package com.example.AutoDocX.service;
 
-import org.springframework.beans.factory.annotation.Autowired;
+import com.example.AutoDocX.service.agent.data.Session;
 import org.springframework.stereotype.Service;
 
 import java.util.Map;
