+++ src/main/java/com/example/AutoDocX/service/agent/DocumentationAgent.java
@@ -1,4 +1,4 @@
-package com.example.AutoDocX.service;
+package com.example.AutoDocX.service.agent;
 
 import com.example.AutoDocX.model.ClonedRepo;
 import com.example.AutoDocX.model.Documentation;
@@ -9,6 +9,12 @@ import com.example.AutoDocX.model.ModelFinishReason;
 import com.example.AutoDocX.parser.model.Graph;
 import com.example.AutoDocX.parser.model.GraphAlgo;
 import com.example.AutoDocX.parser.model.GraphNode;
+import com.example.AutoDocX.service.*;
+import com.example.AutoDocX.service.agent.tools.McpToolKit;
+import com.example.AutoDocX.service.agent.util.MessageBuilder;
+import com.example.AutoDocX.service.agent.data.Session;
+import com.example.AutoDocX.service.agent.tools.ToolExecutionContext;
+import com.example.AutoDocX.service.agent.memory.Memory;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.google.genai.types.Content;
