+++ src/main/java/com/example/AutoDocX/service/agent/memory/MemoryInterface.java
@@ -1,4 +1,4 @@
-package com.example.AutoDocX.service;
+package com.example.AutoDocX.service.agent.memory;
 
 import lombok.Getter;
