+++ src/main/java/com/example/AutoDocX/service/agent/tools/ToolExecutionContext.java
@@ -1,7 +1,9 @@
-package com.example.AutoDocX.service;
+package com.example.AutoDocX.service.agent.tools;
 
 import com.example.AutoDocX.model.ClonedRepo;
 import com.example.AutoDocX.parser.model.Graph;
+import com.example.AutoDocX.service.agent.data.Session;
+import com.example.AutoDocX.service.agent.memory.EpisodicMemory;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
